+++ src/app/(payload)/admin/importMap.js
@@ -20,6 +20,7 @@ import { HorizontalRuleFeatureClient as HorizontalRuleFeatureClient_e70f5e05f09f
 import { BlocksFeatureClient as BlocksFeatureClient_e70f5e05f09f93e00b997edb1ef0c864 } from '@payloadcms/richtext-lexical/client'
 import { FolderTableCell as FolderTableCell_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
 import { FolderField as FolderField_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
+import { CopyTextareaButtons as CopyTextareaButtons_6d01d2997ece014a810a0004aba0d05b } from '@/components/Fields/CopyTextareaButtons'
 import { GenerateDraftsMenuItem as GenerateDraftsMenuItem_1c0cdb264b2de2cda3412196fddee5b6 } from '@/components/GenerateDraftsMenuItem'
 import { LinkToDoc as LinkToDoc_aead06e4cbf6b2620c5c51c9ab283634 } from '@payloadcms/plugin-search/client'
 import { ReindexButton as ReindexButton_aead06e4cbf6b2620c5c51c9ab283634 } from '@payloadcms/plugin-search/client'
@@ -56,6 +57,7 @@ export const importMap = {
   "@payloadcms/richtext-lexical/client#BlocksFeatureClient": BlocksFeatureClient_e70f5e05f09f93e00b997edb1ef0c864,
   "@payloadcms/next/rsc#FolderTableCell": FolderTableCell_f9c02e79a4aed9a3924487c0cd4cafb1,
   "@payloadcms/next/rsc#FolderField": FolderField_f9c02e79a4aed9a3924487c0cd4cafb1,
+  "@/components/Fields/CopyTextareaButtons#CopyTextareaButtons": CopyTextareaButtons_6d01d2997ece014a810a0004aba0d05b,
   "@/components/GenerateDraftsMenuItem#GenerateDraftsMenuItem": GenerateDraftsMenuItem_1c0cdb264b2de2cda3412196fddee5b6,
   "@payloadcms/plugin-search/client#LinkToDoc": LinkToDoc_aead06e4cbf6b2620c5c51c9ab283634,
   "@payloadcms/plugin-search/client#ReindexButton": ReindexButton_aead06e4cbf6b2620c5c51c9ab283634,